blog_platform/requirements.txt
django
djangorestframework
django-filter
Faker
psycopg2-binary 